Femmecare/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
channels==4.1.0
Django==4.2.13
django-cors-headers==4.3.1
django-otp==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
pyotp==2.9.0
sqlparse==0.5.0
typing_extensions==4.12.0
